Convert a list or vector to a batched list of its elements

Batch up a long vector, or list of vectors. For example so they can be passed via a map function to services with length-limited APIs.

Usage

batch_it(x, batches = NULL, proportion = NULL, maximise = FALSE, quiet = TRUE)

batch_it_simple(x, batch_size)

Arguments

x

a vector, or a list flattenable to a vector

batches

numeric. The size (length) of batches to create. Can be a single value or multiple values (see Examples). Should be a whole, positive number, if provided, else NULL.

proportion

numeric. Proportional sizes of batches to be created. For example c(4, 6) will create two batches of approximately 40% and 60% of the length of the target vector (x). When multiple proportion values are provided, these are not repeated. A single proportion less than 1 is repeated as many times as possible to get near to the length of the target vector. For example, a proportion of 0.1 will be treated as a tenth, and batch sizes will be rounded to an integer size nearest to a tenth of the length of x.

maximise

Boolean, FALSE by default. If TRUE, a vector of batch sizes will be partially repeated to fit maximally to the length of the target vector. See examples below.

quiet

Boolean, TRUE by default. Whether to show informative ui_* messages from {usethis}.

batch_size

numeric. The size (length) of batches to create. Should be a single value (see Examples). If supplied as a decimal (<1), it will be interpreted as a proportion of length(x).

Value

All the elements of x batched into a list.

See also

Examples

batch_it(seq(2L, 60L, 2L), 6L)
#> [[1]]
#> [1]  2  4  6  8 10 12
#> 
#> [[2]]
#> [1] 14 16 18 20 22 24
#> 
#> [[3]]
#> [1] 26 28 30 32 34 36
#> 
#> [[4]]
#> [1] 38 40 42 44 46 48
#> 
#> [[5]]
#> [1] 50 52 54 56 58 60
#> 
batch_it(seq(2L, 60L, 2L), proportion = 0.2)
#> [[1]]
#> [1]  2  4  6  8 10 12
#> 
#> [[2]]
#> [1] 14 16 18 20 22 24
#> 
#> [[3]]
#> [1] 26 28 30 32 34 36
#> 
#> [[4]]
#> [1] 38 40 42 44 46 48
#> 
#> [[5]]
#> [1] 50 52 54 56 58 60
#> 

batch_it(1:100, batches = c(20L, 30L, 50L))
#> [[1]]
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20
#> 
#> [[2]]
#>  [1] 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45
#> [26] 46 47 48 49 50
#> 
#> [[3]]
#>  [1]  51  52  53  54  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69
#> [20]  70  71  72  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88
#> [39]  89  90  91  92  93  94  95  96  97  98  99 100
#> 
batch_it(letters, batches = c(4L, 6L))
#> [[1]]
#> [1] "a" "b" "c" "d"
#> 
#> [[2]]
#> [1] "e" "f" "g" "h" "i" "j"
#> 
#> [[3]]
#> [1] "k" "l" "m" "n"
#> 
#> [[4]]
#> [1] "o" "p" "q" "r" "s" "t"
#> 
#> [[5]]
#> [1] "u" "v" "w" "x" "y" "z"
#> 
batch_it(letters, batches = c(4L, 6L), maximise = TRUE)
#> [[1]]
#> [1] "a" "b" "c" "d"
#> 
#> [[2]]
#> [1] "e" "f" "g" "h" "i" "j"
#> 
#> [[3]]
#> [1] "k" "l" "m" "n"
#> 
#> [[4]]
#> [1] "o" "p" "q" "r" "s" "t"
#> 
#> [[5]]
#> [1] "u" "v" "w" "x"
#> 
#> [[6]]
#> [1] "y" "z"
#> 
batch_it(letters, proportion = c(4L, 6L))
#> [[1]]
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j"
#> 
#> [[2]]
#>  [1] "k" "l" "m" "n" "o" "p" "q" "r" "s" "t" "u" "v" "w" "x" "y" "z"
#> 

# ----
as_year <- function(x) {
  lubridate::as_date(
    lubridate::ymd(paste0(x, "-01-01")):
    lubridate::ymd(paste0(x, "-12-31"))
  )
}
month_lengths <- function(year) {
  lubridate::as_date(paste0(year, "-", 1:12, "-01")) |>
    lubridate::days_in_month()
}
batch_it(x = as_year(2022L), batches = month_lengths(2022L))
#> $Jan
#>  [1] "2022-01-01" "2022-01-02" "2022-01-03" "2022-01-04" "2022-01-05"
#>  [6] "2022-01-06" "2022-01-07" "2022-01-08" "2022-01-09" "2022-01-10"
#> [11] "2022-01-11" "2022-01-12" "2022-01-13" "2022-01-14" "2022-01-15"
#> [16] "2022-01-16" "2022-01-17" "2022-01-18" "2022-01-19" "2022-01-20"
#> [21] "2022-01-21" "2022-01-22" "2022-01-23" "2022-01-24" "2022-01-25"
#> [26] "2022-01-26" "2022-01-27" "2022-01-28" "2022-01-29" "2022-01-30"
#> [31] "2022-01-31"
#> 
#> $Feb
#>  [1] "2022-02-01" "2022-02-02" "2022-02-03" "2022-02-04" "2022-02-05"
#>  [6] "2022-02-06" "2022-02-07" "2022-02-08" "2022-02-09" "2022-02-10"
#> [11] "2022-02-11" "2022-02-12" "2022-02-13" "2022-02-14" "2022-02-15"
#> [16] "2022-02-16" "2022-02-17" "2022-02-18" "2022-02-19" "2022-02-20"
#> [21] "2022-02-21" "2022-02-22" "2022-02-23" "2022-02-24" "2022-02-25"
#> [26] "2022-02-26" "2022-02-27" "2022-02-28"
#> 
#> $Mar
#>  [1] "2022-03-01" "2022-03-02" "2022-03-03" "2022-03-04" "2022-03-05"
#>  [6] "2022-03-06" "2022-03-07" "2022-03-08" "2022-03-09" "2022-03-10"
#> [11] "2022-03-11" "2022-03-12" "2022-03-13" "2022-03-14" "2022-03-15"
#> [16] "2022-03-16" "2022-03-17" "2022-03-18" "2022-03-19" "2022-03-20"
#> [21] "2022-03-21" "2022-03-22" "2022-03-23" "2022-03-24" "2022-03-25"
#> [26] "2022-03-26" "2022-03-27" "2022-03-28" "2022-03-29" "2022-03-30"
#> [31] "2022-03-31"
#> 
#> $Apr
#>  [1] "2022-04-01" "2022-04-02" "2022-04-03" "2022-04-04" "2022-04-05"
#>  [6] "2022-04-06" "2022-04-07" "2022-04-08" "2022-04-09" "2022-04-10"
#> [11] "2022-04-11" "2022-04-12" "2022-04-13" "2022-04-14" "2022-04-15"
#> [16] "2022-04-16" "2022-04-17" "2022-04-18" "2022-04-19" "2022-04-20"
#> [21] "2022-04-21" "2022-04-22" "2022-04-23" "2022-04-24" "2022-04-25"
#> [26] "2022-04-26" "2022-04-27" "2022-04-28" "2022-04-29" "2022-04-30"
#> 
#> $May
#>  [1] "2022-05-01" "2022-05-02" "2022-05-03" "2022-05-04" "2022-05-05"
#>  [6] "2022-05-06" "2022-05-07" "2022-05-08" "2022-05-09" "2022-05-10"
#> [11] "2022-05-11" "2022-05-12" "2022-05-13" "2022-05-14" "2022-05-15"
#> [16] "2022-05-16" "2022-05-17" "2022-05-18" "2022-05-19" "2022-05-20"
#> [21] "2022-05-21" "2022-05-22" "2022-05-23" "2022-05-24" "2022-05-25"
#> [26] "2022-05-26" "2022-05-27" "2022-05-28" "2022-05-29" "2022-05-30"
#> [31] "2022-05-31"
#> 
#> $Jun
#>  [1] "2022-06-01" "2022-06-02" "2022-06-03" "2022-06-04" "2022-06-05"
#>  [6] "2022-06-06" "2022-06-07" "2022-06-08" "2022-06-09" "2022-06-10"
#> [11] "2022-06-11" "2022-06-12" "2022-06-13" "2022-06-14" "2022-06-15"
#> [16] "2022-06-16" "2022-06-17" "2022-06-18" "2022-06-19" "2022-06-20"
#> [21] "2022-06-21" "2022-06-22" "2022-06-23" "2022-06-24" "2022-06-25"
#> [26] "2022-06-26" "2022-06-27" "2022-06-28" "2022-06-29" "2022-06-30"
#> 
#> $Jul
#>  [1] "2022-07-01" "2022-07-02" "2022-07-03" "2022-07-04" "2022-07-05"
#>  [6] "2022-07-06" "2022-07-07" "2022-07-08" "2022-07-09" "2022-07-10"
#> [11] "2022-07-11" "2022-07-12" "2022-07-13" "2022-07-14" "2022-07-15"
#> [16] "2022-07-16" "2022-07-17" "2022-07-18" "2022-07-19" "2022-07-20"
#> [21] "2022-07-21" "2022-07-22" "2022-07-23" "2022-07-24" "2022-07-25"
#> [26] "2022-07-26" "2022-07-27" "2022-07-28" "2022-07-29" "2022-07-30"
#> [31] "2022-07-31"
#> 
#> $Aug
#>  [1] "2022-08-01" "2022-08-02" "2022-08-03" "2022-08-04" "2022-08-05"
#>  [6] "2022-08-06" "2022-08-07" "2022-08-08" "2022-08-09" "2022-08-10"
#> [11] "2022-08-11" "2022-08-12" "2022-08-13" "2022-08-14" "2022-08-15"
#> [16] "2022-08-16" "2022-08-17" "2022-08-18" "2022-08-19" "2022-08-20"
#> [21] "2022-08-21" "2022-08-22" "2022-08-23" "2022-08-24" "2022-08-25"
#> [26] "2022-08-26" "2022-08-27" "2022-08-28" "2022-08-29" "2022-08-30"
#> [31] "2022-08-31"
#> 
#> $Sep
#>  [1] "2022-09-01" "2022-09-02" "2022-09-03" "2022-09-04" "2022-09-05"
#>  [6] "2022-09-06" "2022-09-07" "2022-09-08" "2022-09-09" "2022-09-10"
#> [11] "2022-09-11" "2022-09-12" "2022-09-13" "2022-09-14" "2022-09-15"
#> [16] "2022-09-16" "2022-09-17" "2022-09-18" "2022-09-19" "2022-09-20"
#> [21] "2022-09-21" "2022-09-22" "2022-09-23" "2022-09-24" "2022-09-25"
#> [26] "2022-09-26" "2022-09-27" "2022-09-28" "2022-09-29" "2022-09-30"
#> 
#> $Oct
#>  [1] "2022-10-01" "2022-10-02" "2022-10-03" "2022-10-04" "2022-10-05"
#>  [6] "2022-10-06" "2022-10-07" "2022-10-08" "2022-10-09" "2022-10-10"
#> [11] "2022-10-11" "2022-10-12" "2022-10-13" "2022-10-14" "2022-10-15"
#> [16] "2022-10-16" "2022-10-17" "2022-10-18" "2022-10-19" "2022-10-20"
#> [21] "2022-10-21" "2022-10-22" "2022-10-23" "2022-10-24" "2022-10-25"
#> [26] "2022-10-26" "2022-10-27" "2022-10-28" "2022-10-29" "2022-10-30"
#> [31] "2022-10-31"
#> 
#> $Nov
#>  [1] "2022-11-01" "2022-11-02" "2022-11-03" "2022-11-04" "2022-11-05"
#>  [6] "2022-11-06" "2022-11-07" "2022-11-08" "2022-11-09" "2022-11-10"
#> [11] "2022-11-11" "2022-11-12" "2022-11-13" "2022-11-14" "2022-11-15"
#> [16] "2022-11-16" "2022-11-17" "2022-11-18" "2022-11-19" "2022-11-20"
#> [21] "2022-11-21" "2022-11-22" "2022-11-23" "2022-11-24" "2022-11-25"
#> [26] "2022-11-26" "2022-11-27" "2022-11-28" "2022-11-29" "2022-11-30"
#> 
#> $Dec
#>  [1] "2022-12-01" "2022-12-02" "2022-12-03" "2022-12-04" "2022-12-05"
#>  [6] "2022-12-06" "2022-12-07" "2022-12-08" "2022-12-09" "2022-12-10"
#> [11] "2022-12-11" "2022-12-12" "2022-12-13" "2022-12-14" "2022-12-15"
#> [16] "2022-12-16" "2022-12-17" "2022-12-18" "2022-12-19" "2022-12-20"
#> [21] "2022-12-21" "2022-12-22" "2022-12-23" "2022-12-24" "2022-12-25"
#> [26] "2022-12-26" "2022-12-27" "2022-12-28" "2022-12-29" "2022-12-30"
#> [31] "2022-12-31"
#> 

# ----
batch_it_simple(letters, 6L)
#> [[1]]
#> [1] "a" "b" "c" "d" "e" "f"
#> 
#> [[2]]
#> [1] "g" "h" "i" "j" "k" "l"
#> 
#> [[3]]
#> [1] "m" "n" "o" "p" "q" "r"
#> 
#> [[4]]
#> [1] "s" "t" "u" "v" "w" "x"
#> 
#> [[5]]
#> [1] "y" "z"
#> 
batch_it_simple(letters, 0.45)
#> [[1]]
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l"
#> 
#> [[2]]
#>  [1] "m" "n" "o" "p" "q" "r" "s" "t" "u" "v" "w" "x"
#> 
#> [[3]]
#> [1] "y" "z"
#>